Remove content_scripts and use activeTab instead
Chrome store was flagging an issue with the extension having 'Broad
host permissions' whcih was due to matching all urls when running the
content.js script.
To fix, removed 'content_scripts' from the manifext, and instead using
activeTab to inject the content.js script into only the active tab when
the extension is clicked.